requirements.txt
asgiref==3.5.2
attrs==21.4.0
bleach==5.0.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
click-aliases==1.0.1
divio-cli==3.15.6
Django==4.0.5
django-modeltranslation==0.18.2
django-summernote==0.8.20.0
idna==3.3
packaging==21.3
Pillow==9.1.1
pyaml==21.10.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.0
sentry-sdk==1.5.12
six==1.16.0
sqlparse==0.4.2
tabulate==0.8.10
urllib3==1.26.9
uWSGI==2.0.20
webencodings==0.5.1
